APIs/twilio.com/twilio_voice_v1/1.55.0/python-aiohttp-client/tests/test_voice_v1_source_ip_mapping_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.list_source_ip_mapping_response import ListSourceIpMappingResponse
from openapi_server.models.voice_v1_source_ip_mapping import VoiceV1SourceIpMapping


pytestmark = pytest.mark.asyncio

@pytest.mark.skip("application/x-www-form-urlencoded not supported by Connexion")
async def test_create_source_ip_mapping(client):
    """Test case for create_source_ip_mapping

    
    """
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/x-www-form-urlencoded',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    data = {
        'ip_record_sid': 'ip_record_sid_example',
        'sip_domain_sid': 'sip_domain_sid_example'
        }
    response = await client.request(
        method='POST',
        path='/v1/SourceIpMappings',
        headers=headers,
        data=data,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_delete_source_ip_mapping(client):
    """Test case for delete_source_ip_mapping

    
    """
    headers = { 
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='DELETE',
        path='/v1/SourceIpMappings/{sid}'.format(sid='sid_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_fetch_source_ip_mapping(client):
    """Test case for fetch_source_ip_mapping

    
    """
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='GET',
        path='/v1/SourceIpMappings/{sid}'.format(sid='sid_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_list_source_ip_mapping(client):
    """Test case for list_source_ip_mapping

    
    """
    params = [('PageSize', 56),
                    ('Page', 56),
                    ('PageToken', 'page_token_example')]
    headers = { 
        'Accept': 'application/json',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    response = await client.request(
        method='GET',
        path='/v1/SourceIpMappings',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

@pytest.mark.skip("application/x-www-form-urlencoded not supported by Connexion")
async def test_update_source_ip_mapping(client):
    """Test case for update_source_ip_mapping

    
    """
    headers = { 
        'Accept': 'application/json',
        'Content-Type': 'application/x-www-form-urlencoded',
        'Authorization': 'BasicZm9vOmJhcg==',
    }
    data = {
        'sip_domain_sid': 'sip_domain_sid_example'
        }
    response = await client.request(
        method='POST',
        path='/v1/SourceIpMappings/{sid}'.format(sid='sid_example'),
        headers=headers,
        data=data,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

